requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
contourpy==1.0.6
cycler==0.11.0
filelock==3.9.0
fonttools==4.38.0
idna==3.4
imageio==2.22.2
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.2
networkx==2.8.7
numpy==1.23.3
opencv-python==4.5.5.64
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
threadpoolctl==3.1.0
tifffile==2022.10.10
torch==1.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.64.1
typing-extensions==4.3.0
urllib3==1.26.12
